+++ frontend/e2e-tests/MoEmployeeCreate.js
@@ -53,18 +53,8 @@ const itSystemInput = dialog.find('.input-itSystem input[data-vv-as="Kontonavn"]
 // Manager
 const parentManagerInput = dialog.find('.unit-manager input[data-vv-as="Angiv enhed"]')
 
-const addressManagerTypeSelect = dialog.find('.address-manager select[data-vv-as="Adressetype"]')
-const addressManagerTypeOption = addressManagerTypeSelect.find('option')
-
-const addressManagerInput = dialog.find('.address-manager .v-autocomplete[data-vv-as="Adresse"]')
-const addressManagerItem = addressManagerInput.find('.v-autocomplete-list-item label')
-
 const managerTypeSelect = dialog.find('.select-manager select[data-vv-as="Ledertype"]')
 const managerTypeOption = managerTypeSelect.find('option')
-
-const addressManagerMany = dialog.find('.address-manager button')
-
-const addressManagerInputEmail = dialog.find('input[data-vv-as="Email"]')
 
 const levelManagerSelect = dialog.find('.select-manager select[data-vv-as="Lederniveau"]')
 const levelManagerOption = levelManagerSelect.find('option')
@@ -162,23 +152,6 @@ test('Workflow: create employee', async t => {
 
     .click(parentManagerInput)
     .click(dialog.find('.unit-manager .tree-anchor'))
-
-    .click(addressManagerTypeSelect)
-    .click(addressManagerTypeOption.nth(1))
-
-    .click(addressManagerInput)
-    .typeText(addressManagerInput.find('input'), 'baa')
-    .expect(addressManagerItem.withText('Bålvej').visible)
-    .ok()
-    .pressKey('down down down down down down enter')
-    .expect(addressManagerInput.find('input').value)
-    .eql('Bålvej 1, 9800 Hjørring')
-
-    .click(addressManagerMany)
-    .pressKey('tab tab tab tab tab down down enter')
-
-    .click(addressManagerInputEmail)
-    .typeText(dialog.find('input[data-vv-as="Email"]'), 'hest@test.dk')
 
     .click(managerTypeSelect)
     .click(managerTypeOption.withText('Direktør'))