+++ frontend/src/components/MoEntry/MoManagerEntry.vue
@@ -20,16 +20,6 @@
         :validity="entry.validity"
       />
 
-      <mo-add-many
-        class="address-manager"
-        v-model="entry.address"
-        :entry-component="managerAddressEntry"
-        :label="$t('input_fields.manager_address_type')"
-        validity-hidden
-        has-initial-entry
-        small-buttons
-      />
-
     <div class="form-row select-manager">
       <mo-facet-picker
         facet="manager_type"
@@ -71,7 +61,6 @@ import MoOrganisationUnitPicker from '@/components/MoPicker/MoOrganisationUnitPi
 import MoFacetPicker from '@/components/MoPicker/MoFacetPicker'
 import MoAddMany from '@/components/MoAddMany/MoAddMany'
 import MoEmployeePicker from '@/components/MoPicker/MoEmployeePicker'
-import { MoManagerAddressEntry } from '@/components/MoEntry'
 import MoEntryBase from './MoEntryBase.js'
 import OrgUnitValidity from '@/mixins/OrgUnitValidity'
 
@@ -129,29 +118,6 @@ export default {
         },
 
         template: `<div class="form-row"><mo-facet-picker facet="responsibility" v-model="val" required/></div>`
-      }
-    },
-
-    /**
-     * Adds the managerAddressEntry template to the add many component.
-     */
-    managerAddressEntry () {
-      return {
-        components: {
-          MoManagerAddressEntry
-        },
-
-        props: {
-          value: [Object, Array]
-        },
-
-        data () {
-          return {
-            val: this.value
-          }
-        },
-
-        template: `<mo-manager-address-entry v-model="val" validity-hidden required/>`
       }
     }
   },