+++ frontend/src/components/MoEntry/index.js
@@ -4,10 +4,9 @@ import MoEmployeeAddressEntry from './MoEmployeeAddressEntry.vue'
 import MoEngagementEntry from './MoEngagementEntry.vue'
 import MoItSystemEntry from './MoItSystemEntry.vue'
 import MoLeaveEntry from './MoLeaveEntry.vue'
-import MoManagerAddressEntry from './MoManagerAddressEntry.vue'
 import MoManagerEntry from './MoManagerEntry.vue'
 import MoOrganisationUnitEntry from './MoOrganisationUnitEntry.vue'
 import MoOrgUnitAddressEntry from './MoOrgUnitAddressEntry.vue'
 import MoRoleEntry from './MoRoleEntry.vue'
 
-export { MoAddressEntry, MoAssociationEntry, MoEmployeeAddressEntry, MoEngagementEntry, MoItSystemEntry, MoLeaveEntry, MoManagerAddressEntry, MoManagerEntry, MoOrganisationUnitEntry, MoOrgUnitAddressEntry, MoRoleEntry }
+export { MoAddressEntry, MoAssociationEntry, MoEmployeeAddressEntry, MoEngagementEntry, MoItSystemEntry, MoLeaveEntry, MoManagerEntry, MoOrganisationUnitEntry, MoOrgUnitAddressEntry, MoRoleEntry }